shill: Move NotifyDarkResumeScanResultsReceived to WakeOnWiFi.
The metrics is only valuable for WakeOnWiFi, so it makes more sense to
have the logic in WakeOnWiFi. This can also unexport IsDarkResume
function of WakeOnWiFi. It doesn't seem correct for WiFi to rely
on WakeOnWiFi::IsDarkResume. If WiFi object needs the information,
it should track it by itself.

BUG=b:172208175
TEST=unit tests

Change-Id: I889609ca0e3dc58057232eda85f219fcf956e21b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2763935
Tested-by: Yen-lin Lai <yenlinlai@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>