hermes: Convert SMDP domain name to lowercase
QR codes from a certain vendor encode the SMDP address in uppercase.
While this is not supposed to impact the servers response to
initiateAuthenticate, it does in practice. A request to SMDP.COM results
in a http 400 error, but a request to smdp.com results in an install.
This CL converts smdp addresses in activation codes to lowercase.

BUG=b:183032912
TEST=modem esim install '1$SMDP.COM$ACTIVATION-CODE'

Change-Id: Id7720f2f17e0b33b0a07f73508463dcf6fab6f3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2769285
Tested-by: Pavan Holla <pholla@google.com>
Auto-Submit: Pavan Holla <pholla@google.com>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>