shill: Add WakeOnWiFiSupported property.
After CL:2763546, the callers, e.g. UI, will need to check the existence
of WakeOnWiFi properties on WiFi device to know if WakeOnWiFi is
supported, which is not very straightforward.

Add a supported boolean property on WiFi device for caller to check more
easily.

Also, this will replace the old EnabledFeature="not_supported" condition,
and we can remove kWakeOnWiFiFeaturesEnabledNotSupported as it is no
longer used.

BUG=b:172208175
TEST=unit tests
TEST=checked with dbus-send

Change-Id: Id2275b1624332ad5103457f4ffd13372ada74fc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2775058
Tested-by: Yen-lin Lai <yenlinlai@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>