modemfwd: add variant filtering to helper_manifest
Add a new (repeated) field 'variant' in the helper_manifest.prototxt
configuration file in order to filter out the helpers corresponding to
non-existing modems depending on the 'variant' of a unibuild.

This avoids trying to reset a modem in 'wedged'/recovery detection while
it only exists on another type of machines.

BUG=b:183180499
BUG=b:176138160
TEST=Use an updated helper_manifest.prototxt for Zork and see that we
are not seeing spurious modem reboot on vilboz AND vilboz360.

Change-Id: Iaf2cd456925008dde800ca21b3f575614ba939cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2773127
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Henry Sun <henrysun@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Henry Sun <henrysun@google.com>