shill: Cellular: Set the (friendly) Name property correctly
Currently the friendly name is cleared in SetDevice.
For the primary Service, the name will get overwritten when
Cellular::UpdateServingOperator gets called, however in
secondary services this always results in a generic name.

This CL sets a generic friendly name in the CellularService
constructor and only overrides it in
Cellular::UpdateServingOperator if a friendly name is available.

This also loads the persisted friendly name so that secondary
services use the last saved friendly name instead of a generic one.

BUG=b:182488343
TEST=unit tests + manual testing

Change-Id: I25095983baa24b2c33e014a35d5f8924a4b8dbc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2775872
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>