usb_bouncer: Add unblocked_terms.txt.
This serves as a presubmit check for non-inclusive words in code.

Unfortunately the user of the word 'sl*ve' is in an MS_ mount flag so
it's unlikely we'll be able to remove it. But this prevents anything
else from creeping in.

BUG=None
TEST=Upload.

Change-Id: I276a63e779c3866c379acbbe382a6a5343397106
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2776072
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>