vm_tools: Remount the 'mnt_concierge' bind-mount as 'noexec'.
We're starting a project to report W+X mounts (which violate verified
boot) and this mount would trigger it.

While ns bind mounts are unlikely to be used to bypass verified boot,
we should not have W+X mounts on the system. Fix this by making it
'noexec'. This file is only ever used as the target of a setns(2)
system call.

BUG=chromium:1190277
TEST=Deploy, start session, session starts correctly.

Change-Id: I2365ef3cdf6e5542f93d9f74d06697f4cd5f8f60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2776074
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>