typecd: Add peripheral data access variable
To prepare for the change in policy to rely on a Chrome setting to
determine mode switches, introduce a peripheral data access variable
which tracks the state of that setting. Also update the mode entry logic
to account for this variable while making mode switching decisions.

The variable is currently set to true, so this should not change the
behaviour of typecd mode switching.

Also add some unit tests to test the behaviour of this setting (by
simulating its change manually) for the following scenarios:
- When the flag is set to false, don't switch modes on login.
- When the flag is set to true *after* hotplug, make sure a switch
  occurs on the next login.

BUG=b:182835509
TEST=- Run all unit tests.
     - Run typecd.ModeSwitch, typecd.ModeReboot and typecd.ModeCrash for
     10 iterations each.

Change-Id: I9e878208d939f8c3e43d3723e8c8139dfa73d2e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2775889
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>