Revert "power: udev: Set DLPM based initial value and device"
This reverts commit 221680671b2835697c76a04f4714047d1f777e7a.

Set DLPM in kernel based on horkage.
On Fizz, not only disable DLPM for Transcend SSD, but for all SSD.
Transcend was only qualified for Fizz chromebox.

BUG=b:180139522
TEST=Check device DLPM is "max_performance".
Check that setting it to "min_power" fails, thus is can not be enabled
on the SSD.

Cq-Depend: chromium:2776399,chromium:2787456
Change-Id: I9694d82ab8313d809277b6b44fa68d7f5fb13269
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2773451
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Rajat Jain <rajatja@google.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>