vm_tools: p9: Allow fchown and fchmod
The 9p server is not the right place to make these kinds of policy
decisions.  Instead we should rely on the sandboxing environment to
enforce what the server is allowed to do.

This change by itself is not enough to fully support fchown however. For
crostini, the 9p server runs as the chronos user with no other
capabilities so attempting to the change to uid/gid of a file to
anything other than chronos will still fail. However, fchmod and no-op
uid/gid changes should succeed now.

BUG=chromium:1082345,chromium:920262,chromium:912599
TEST=Moving a file from the container to a shared folder doesn't
     complain about fchmod failing.  `chown 1000:1000` on a shared
     file/directory also does not fail.

Cq-Depend: chromium:2784163
Change-Id: Ibad83a1ef182c6f2962d7e180d0e9798b7d73a9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2784361
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>