tpm_manager: disable DA mitigation mechanism for TPM2.0.
We don't need DA mitigation to protect the device while the DA
mitigation itself has been causing real issues. This is done so TPM2.0
devices like cr50 and generic TPM2.0 can be DA-lockout-free.

For testing on real devices, an already owned device is deployed with
the new tpm manager, reboot, and writing a nv index with wrong auth
doesn't increase the DA.
Likewise, for an unowned device, taking TPM ownership with the new tpm
manager, writing a nv index with wrong auth, and the DA counter doesn't
increase either.

Note that the on-device test doesn't test the wrong lockout auth; we
assume the TPM implementation is correct and we already make sure
lockout recovery time is set to zero in unittests.

BUG=b:183477880 # disable DA at userland side.
BUG=b:178365982 # disable DA on cr50
TEST=unittests.
TEST=See above for real device test.

Cq-Depend: chromium:2784306
Change-Id: I015d6533c71eb2711b3f2c03258e1d5ae74f1075
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2784366
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Leo Lai <cylai@google.com>
Tested-by: Leo Lai <cylai@google.com>