patchpanel: allow ARC traffic to exit any interface
Starting from m89, the datapath was modified so that ARC traffic routed
to a given physical interface had to originate from the corresponding
virtual interface. i.e ARC traffic routed through wlan0 could only come
from arc_wlan0, etc.

This created a regression for some multihoming scenarios when the
device is connected to the same logical network through two network
interfaces and ARC is sending traffic to a peer on the LAN.

To fix this regression, this patch relaxes the iptables rules
controlling which physical interfaces ARC traffic can be forwarded to
and now allows any output interface as it was the case before m89.

Note that this does not impact the connection pinning feature shipped in
m89. This is because iptables FORWARD chain in the filter table already
add an ACCEPT-all rule for any traffic already known to conntrack.

BUG=b:182594063
TEST=unit test. Flashed hatch-arc-r, connected hatch to the same network
over wifi and ethernet, checked that ARC can reach a local name server
through both networks.

Change-Id: Ib9273d83e7ec57da279dc97274a673b597713627
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2784364
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>