shill: Cellular: Generate errors when SetInhibited fails.
This CL includes 3 changes:
1. This fixes a bug where |error| was not getting set when
SetInhibited failed (e.g. when Cellular is in an unexpected
state). This was causing the DBus operation to succeed, even
though Inhibited was not set.

2. When no SIMs are available, ModemManager fails to start,
leaving Cellular.capability_state_ in CapabilityStatrted
instead of ModemStarted, causing SetInhibited to fail. We now
also allow SetInhibited to proceed if modem_state_ is failed.

3. This also moves some logic to fetch the MM.Device (UID)
property to SetInhibited instead of StartModem to simplify
StartModem and to ensure that |uid| is not stale.

BUG=b:172077101
TEST=unit tests, manual testing

Change-Id: I200a249e30840ee21a5cfeb1851be0a439514039
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2785960
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>