shill: cellular: add support for setting attach APN through the UI
When receiving a (manual) APN setting from the UI, save it and restore
it properly (ie including whether to use it as an attach APN).

When we get an attach APN setting, detach / re-attach to the network in
order to use the new settings.

BUG=chromium:1191189
TEST=unit-tests
TEST=On Boten(L850) and Lazor, run w/o and with #cellular-use-attach-apn
tast run dut '("group:cellular")'
TEST=manual, on Boten (L850-GL) force attach APN on Swisscom (CH) and
verify it still connects (and the logs show the detach)
TEST=manual, on PomPom (in JP) enter manually the proper attach APN as
above.

Change-Id: I2a25954d18f51063f6462217191bfbbf5622555b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2784693
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>