typecd: Add a skeleton ChromeFeaturesService client
Add setup code to register a proxy with the ChromeFeaturesService in
Chrome. This can be used to get the values of device settings which the
daemon is interested in.

While we are here, give PortManager a pointer to the
ChromeFeaturesServiceClient object, so that it can call it if it needs
to.

BUG=b:182835509
TEST=- Unit tests pass
     - typec.ModeSwitch, typec.ModeReboot pass 10/10 times each.

Change-Id: Ied08ec16a2a9b6c443006bfcecf0138201613d57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2792696
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>