patchpanel: simplify PREROUTING apply_vpn_mark
PREROUTING Fwmark routing rules for VPNs are now always recreated
when a VPN connection is established. This allows to systematically
flush the apply_vpn_mark chain when the VPN disconnects and removes the
need to remember the interface index of the VPN virtual interface at
connection time.

BUG=b:183679000
BUG=b:161060333
TEST=unit tests.

Change-Id: I6a757fdd6239d7d77cd02129498ab474f3e8f173
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2791265
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>