secanomalyd: Add init mount namespace W+X mount metrics.
In order to minimize load to the crash server, we want to have an UMA
entry for every anomaly that we'd eventually report as a crash. This
way we'd be able to estimate pretty accurately how many crashes we'd be
uploading to the crash server.

The first anomaly that secanomalyd will detect is writable and
executable mounts present in the init mount namespace. These anomalies
will be reported to UMA as bucket 0 of the histogram
"ChromeOS.SecurityAnomaly". This bucket will be named
"Mount.InitNS.WX".

Reporting will initially occur only in verified mode, since devices in
dev mode can have unexpected W+X mounts.

BUG=chromium:1197143
TEST=Deploy to device, run with --dev.
TEST=See / and /usr/local listed as W+X mounts (in dev mode.)
TEST=Open chrome:histograms, see two samples reported for bucket 0.

Cq-Depend: chromium:2815442
Change-Id: Id0be8d46aff89f6a23104749a3178f8b7bf9cd09
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2815400
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>