typecd: Add PartnerType metrics reporting
Add the enum definitions and method to report the detected partner type.
Also, move around one of the PD VDO constants (for DP Sink support) so
that it is accessible to the Partner class too.

The CL to add the enum to the metrics histograms XML file is here:
crrev.com/c/2800428

Also add some unit tests to check some of the PartnerType parsing.

BUG=b:164522182
TEST=-Run unit tests.
     -Connect Gatkex, TBT3 Titan Ridge dock and TBT3 Alpine Ridge docks
      and verify the histogram values on chrome://histograms.

Change-Id: Ie336eb65f3cb4bab5237ba4fad8401f5613b8924
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2795897
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>