cryptohome: change ACLs of the user directory in vault
Changing the ACLs in [1] I got confused which direction the permissions
are propagated for bind mounts. This CL makes the vault's user directory
to have the proper ACLs so they are propagated down to the mount points.
It also appeared to be much simpler than I anticipated, though for all
the wrong reasons...

[1] https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2755367

BUG=chromium:1169874
TEST=build, unittest, load on device

Change-Id: I13936d7f5a7c0da08b46b15f4334b1a553f8a7de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2796913
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>