metrics: Set correct permissions on structured metrics event files.
After an events file is created, we want it to have permissions 660.
Currently we are opening the file with 666 permissions, but these are
affected by the umask of whatever process is using this code. This CL
makes two changes:

1. open file with 600 permissions, rather than 666.
2. do an explicit chmod to 660 after writing.

BUG=chromium:1148168
TEST=recording files with a non-022 umask on DUT works correctly.

Disable-Recycled-Builds: test-failures
Change-Id: I71af8c79d8919a72a2aad3c4887d037d242d408c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2796763
Auto-Submit: Tony Yeoman <tby@chromium.org>
Tested-by: Tony Yeoman <tby@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>