chromeos-config: run fallback commands with I_AM_CROS_CONFIG=1
Add a new environment variable when fallback commands are run to
generate configs for pre-unibuild devices, I_AM_CROS_CONFIG=1.

This will allow programs (in particular, I'm speaking of mosys) to
detect that a command is being run as a result of fallback config
generation, and thus allow the command to hide or disable itself
unless it's being used by cros_config.

The usefulness of this is to help prevent further usages of deprecated
commands which have equivalents in cros_config.  For example, this
would allow us to disable "mosys psu type" or "mosys platform brand"
to everyone but cros_config.

BUG=b:184257014
TEST=updated unit tests pass

Change-Id: Id0f0114ade84562696208710d10647a8851ec239
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2799817
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>