ippusb_bridge: Internally release unused USB interfaces
ippusb_bridge claims all the relevant IPP-USB interfaces when it
starts up.  It internally tracks which of these are available, but
doesn't release them until the program exits.  This causes the printer
to be unavailable to other programs (such as cupsd) even if scanning or
printing through IPP-over-USB is not active.

To allow sharing with other programs, call libusb release and claim
functions when moving interfaces in and out of the internal available
pool.  cupsd can now claim interfaces as long as they are not actively
in use, and ippusb_bridge will transparently reclaim them as long as
cupsd is not actively using them when a scan starts.

This introduces a new failure mode to ippusb_bridge: Previously the
interfaces were always available once the program had started, but now
they might be claimed when a scan job starts.  This should only happen
if the user is trying to print and scan simultaneously and is the same
problem that could previous happen with ippusb_manager under the same
circumstances.  A failure to reclaim an interface will leave that
interface in the free pool so that a subsequent retry should succeed
once the conflicting job finishes.

BUG=b:182210048
TEST=Print and scan to XP-7100

Change-Id: I913468e684e1aaffdc3b99e37bce509ff9045436
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2799472
Reviewed-by: Kalvin Lee <kdlee@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>