debugd: Do not log an error when bug report backup doesn't exist
Session Manager deletes the ARC bug report backup on every new
session and that causes an error to be logged in debugd because
the backup doesn't always exist on the disk. With this change we'll
first look for the backup file before attempting to delete it.

BUG=None
TEST=deploy debugd to machine, restart session, and verify error is
not loggged.

Change-Id: I98ca55e48a7868214489d29533116517c6a69fa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2800562
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>