secanomalyd: Add command-line options to control behaviour.
This CL adds two command-line options to control secanomalyd behaviour.

First, we've been asked to only report anomalies to the crash server
after getting an idea of the volume of anomalies that we'd be
reporting. In order to do that this CL adds a command-line flag that
controls whether or not to report anomalies to the crash server. This
flag will be disabled for the time being.

Second, we don't really want to report writable+executable mounts as
anomalies while in dev mode, because these mounts are not unexpected
(are not anomalies, essentially) while in dev mode. Reporting them
would pollute our data. However, we need to be able to test this
behaviour in dev mode because that is our, well, development
environment. So add a flag that can be used by developers to test
secanomalyd behaviour in dev mode. This flag is expected to be used
only during local development.

BUG=chromium:1195377
TEST=Deploy, and test:
TEST=secanomalyd --v=1 shows W+X mounts.
TEST=secanomalyd --v=1 --generate_reports attempts to generate reports
TEST=but skips because of cros_debug=1.
TEST=secanomalyd --v=1 --generate_reports --dev attempts to generate
TEST=report.

Change-Id: I8c751f09f8a20f9a877668c96ea4a90af81f6628
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2802013
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>