typecd: Keep track of active state on last mode entry.
If a peripheral entered DP mode when PeripheralDataAccess was not
enabled, and the setting was then later enabled, then when the
Chromebook next unlocked, a mode switch would be triggered. This would
abruptly remove all USB devices (including storage devices) which were
on the peripheral.

To avoid this, keep track of the user state when a ModeEntry was last
performed on the port. Then, when we unlock, if a device is in DP mode,
but that mode was entered during an unlocked state, we won't change
modes.

Add a unit test to check for this case.

BUG=b:182835509
TEST=- Unit tests all pass.

Change-Id: I02de1b39da06c0279a2e63c30647a9ea16ebb72c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2803451
Tested-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>