shill: wifi: disable connecting to WPA3 service on non-WPA3 device
Not all WiFi drivers/chips support WPA3, and we're gradually phasing in
support even for those that do. Normally, a WPA3-only Service would
never be successfully configured and set to autoconnectable on a system
that doesn't support it, but things like WiFi Sync mean it's possible
for a Service to be carried from another system that does.

Make WPA3-only services only connectable when the underlying Device
supports it.

BUG=b:184395912
TEST=unit tests
TEST=suite:wifi_matfunc
TEST=WIP local test, with both WPA2/3-mixed and WPA3-pure BSSs in the
     same network

Change-Id: I4fc75a65964185985c9156fd12cb2e5af5bf2a2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2803553
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>