shill: vpn: fix always-on VPN remaining service.
In a Profile, the service defined as always-on VPN is persisted using
its "storage identifier". When a service is created, set as always-on,
then removed, the configuration stays. If a service is created leading
to the same storage identifier, it will be considered as always-on even
if it is not the service previously set, it is a bug.

To avoid the situation, the always-on VPN service setting is cleared
from the profile when the service is removed.

BUG=b:184564123
TEST=shill unit tests; tast; manual testing

Change-Id: I4ac705adc556c60b7bd9dbeaf4a2ad9f881ed093
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2813922
Tested-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Damien Dejean <damiendejean@google.com>