shill: Remove connection_id from service
The connection id of a service is calculated based on the MAC address of
the gateway of this connection. Currently, there are two places may use
this property: 1) for "Network.Shill.ServicesOnSameNetwork" metric, which
seems not really in use now; 2) exposed as a property of Service in
D-Bus API, while this is not mentioned in service-api.txt, and the
kConnectionIdProperty constant is not used in both CrOS and Chrome
codebase. Above all, it should be safe to remove this property.

Since the MAC address of the gateway is provided by shill::LinkMonitor
and the latter is going to be deprecated, this patch removes the
connection id property and the corresponding logic.

BUG=b:147256664
TEST=unit

Cq-Depend: chromium:2835140
Change-Id: Ibce02578418336f0a85a14256ff1b739d40efc70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2814180
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>