arc: libvda: Adapt video.mojom to use the new Mojo types.
This CL adapts all Mojo bindings that were not adapted yet in the
previous CLs which updated the encoder and decoder. All Mojo bindings
now use the new Mojo Remote and Receiver types, and are fully in sync
with Chrome.

Note: The updated bindings are currently not used by libvda, but we
should update them anyway to make sure Chrome and ARCVM are in sync.

BUG=b:174967467
TEST=tast run DUT arc.VideoEncodeAccel.h264_192p_i420_vm
     tast run DUT arc.VideoDecodeAccel.h264_vm

Change-Id: Ib172b170f0b3dbfd85a5bab02d67505fc1206223
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2814109
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Alex Lau <alexlau@chromium.org>