system-proxy: Replace loop.RunOnce() in tests
Some system-proxy tests are complex and require stopping the message
loop in the middle of execution to inspect the current state and
simulate events like a connection job getting canceled before the
proxy resolution job for that connection is resolved.

In order to test the corner cases, test cases run tasks in the message
loop one-by-one.

One test has been reported to fail due to the connection request not
being processed after loop.RunOnce() was called.

This CL fixes the flakiness in the tests by running the loop until a
connection request is processed.

BUG=chromium:1197142
TEST=cros_workon_make --board=endeavour system-proxy --test

Change-Id: I695c920696f920b6fadbfc08ba08e867a22ed897
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2817758
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>