debugd: Run "lspci -vvvnn" command as root.
The command "lspci -vvvnn" runs under a user the does not have enough
permissions to access the PCI device capabilities. Run the command as
root to be able to get the capabilities in the feedback reports.

BUG=b:168563469
TEST=emerge-hatch chromeos-base/debugd
Ran the test feedback.SysInfoPII and printed the output it shows
"lspci_verbose" with capabilities information.

Change-Id: I1f1a84aee1844cf7c1ab45bab399939aacd3fa91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2819482
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>