arc: Run arc-boot-continue as a task
Run arc-boot-continue as a task instead of a pre-start script.
A task can also block the triggering event.
Also, it's the main process of the job so the "stop on" condition
properly works.

A new upstart job arcpp-post-login-services is introduced to control
other jobs' lifetimes.

BUG=chromium:1195703
TEST=tast run ui.LoginPerf

Cq-Depend: chromium:2820253
Change-Id: I40c300890bbffcd8fbc5e5ce5c8e5e76985030c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2821079
Tested-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Ereth McKnight-MacNeil <ereth@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Ryo Hashimoto <hashimoto@chromium.org>