crash: Always include 'channel' in reports.
It's useful for data purposes to always include the channel (i.e.
stable/beta/dev/canary/testimage). Sometimes we can get this from the
lsb-release, and sometimes chrome populates the "channel" field,
but it's annoying to distinguish testimage from the rest using just
lsb-release and we need this for non-chrome crashes, too.

Add a field that just has the channel to simplify.

BUG=b:185155560
TEST=deploy to my DUT, generate a crash, verify channel is present.

Change-Id: I07424b5da9867906a7f5d89c47cfc220f61c6f13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2823209
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Michael Irani <michaelirani@google.com>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>