shill: disable IPv6 routing rules for secondary Cellular networks
This patch changes Shill routing setup such that IPv6 routing rules are
always omitted for a Cellular connection that is not the primary
physical connection.

This prevents applications from accidentally using the Cellular
connection with IPv6 traffic if the primary physical network is an
IPv4-only network.

BUG=b:180521518
TEST=Flashed atlas, checked with 2 dual stack networks of different
technologies that IPv6 routing is disabled on the network that is not
the default network (with a local modification to use WiFi instead of
Cellular).

Change-Id: Ia8b04c1b4f549470191c25aa1edf3e2ea4202e36
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2823752
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>