shill: vpn: Check OpenVPNManagementServer state before calling
A crash may happen if the default physical service change event comes
after OpenVPNDriver::ConnectAsync() is called and before the driver
actually starts the client and the management server. The reason is that
some functions in OpenVPNManagementServer(e.g., Restart()) assume it is
called after Start() so it does not check whether |sockets_| is nullptr.

So this patch adds a check before invoking those functions in
OpenVPNDriver::OnDefaultPhysicalServiceEvent(). It should be safe to
ignore this event if the management server is not started. See the added
comment for more details.

BUG=b:185115202
TEST=unit_tests
TEST=manually triggered the default physical changed events by removing
  or plugging in Ethernet dongle. OpenVPN service worked as expected.

Change-Id: I810734fca5a82dfed6f84ac82b92f7783dfac0e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2823748
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>