shill: dbus-client rely on address instead of method
This patch changes the introspection of the IPConfig object from using
of the Method property to verifying the address and family using the
IPAddress class.

1. Shill is not correctly populating the Method property for at least
some types of VPNs.
2. This patch address a clean-up comment to not rely on the Method
property.

The first issue was resulting in the default network dns-proxy getting
no name server configuration when a VPN was turned on.

BUG=b:185428856
BUG=b:174802586
BUG=b:173065426
TEST=units
TEST=manual on octpus vusing cros test l2tp strongswan vpn

Change-Id: Icae9b8febeb5cc6187baabbbbcb4a272ea7b55b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2825557
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Garrick Evans <garrick@chromium.org>