chromeos-config: rename "ash flags" -> "ash switches"
Apparently "flags" is used to refer to the settings in chrome://flags,
whereas the chrome world calls the command line options as "switches".

Official docs:
https://chromium.googlesource.com/chromium/src/+/HEAD/docs/configuration.md

Clean it up so it's less confusing.

Did not touch "extra-ash-flags" for now due to the external boxster
dependency.

BUG=none
TEST=unit tests

Change-Id: Ia9fa7dde075fdbaedaad973bbbe47b626d437fef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2827245
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>