shill: Handle neighbor reachability event in WiFi class
This patch moves the handler for neighbor reachability events from
Device class into WiFi class, since we are sending UMA metrics only for
WiFi, and will need to listen to this event to take actions only in
WiFi. Also rename the function for the following changes since we will
need both of the two types.

This patch is just a refactor and should have no logic change.

BUG=b:147256664
TEST=unit_tests

Change-Id: I92def53a80538ca5215e6765ce197adf139c8667
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2846360
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>