arc: libvda: decode: move VDA connection to its own struct
We are going to make Sessions refer to their VDA connection by Rc
instead of references. This means both the VDA instance and each session
will need to own its own Rc to the connection, so move the connection to
its own struct which destructor closes the connection to VDA established
in the constructor.

BUG=b:169295147
TEST=Video plays correctly on hatch-arc-r.

Change-Id: Id17126119cb0550139bd174ed069288da04ddfbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2838712
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>