shill: device: persist Enabled state more readily
It's possible for a Device to be disabled non-persistently for $reasons,
but then for it to be disabled persistently for other reasons. The
latter should persist to the profile storage.

Example: a newly-constructed Device starts up enabled_==false but
enabled_persistent_==true. The only way to force such a device to stay
Disabled is to toggle it on and back off.

A few test cases seem to be coded to existing behavior, rather than what
behavior actually makes sense, so I tweaked and extended them.

BUG=none
TEST=unit tests

Change-Id: I69300bcb2e6d2770312da3d97d9b5cef98bfe25a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2965558
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>