shill: KeyValueStore::Lookup() returns value instead of reference
For non-arithmetic types, Lookup() accepts a |default_value| which is a
const reference. It could be dangerous if we return a reference to this
|default_value|, since this input reference could bind to a temporary
object, and it is easy to be misused, e.g.:
  const std::string& a = kvs.Lookup<std::string>(name, "");
In the above example, after this expression, |a| will becomes a dangling
reference, and this problem cannot be detected by the compiler. Changing
to returning a value could avoid this, with the cost that we may need
some additional copies in some cases.

Another solution could be adding another overload function which takes
an rvalue reference as the input and returns a value, but:
- It's confusing that the same function may return either reference or
  value;
- Currently, we only call this for std::string. Returning a copy for
  std::string should be acceptable in our cases;
- It cannot avoid copy totally so if a copy is unacceptable, other
  functions (Contains() and Get()) can and should be used.

BUG=None
TEST=unit_tests

Change-Id: Ie80be4ea0dd4a7d4fbf903b4f764410229dcbecd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2841886
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>